account refactor
Former-commit-id: b12026b3fa07ed2df418deac7e9c400a98e19830
Former-commit-id: aa12c1c4fcdabb0168b4c2755fc45525edfc5fb0

+++ core/domain/interface/member/account.go
@@ -105,8 +105,8 @@ type (
 		// 退款
 		Refund(accountKind int, title string, amount float32, outerNo string, remark string) error
 
-		// 充值
-		Charge(account int32, title string, amount float32, outerNo string, remark string) error
+		// 充值,金额放大100倍
+		Charge(account int32, title string, amount int, outerNo string, remark string) error
 
 		// 客服调整
 		Adjust(account int, title string, amount float32, remark string, relateUser int64) error
